:art: Refactor SearchBar.js
