--- src/components/SearchBar/SearchBar.css
@@ -1,23 +1,0 @@
-#search-bar {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    margin-bottom: 1rem;
-  }
-  
-  #search-bar .text-filter {
-    flex-grow: 1;
-    flex-shrink: 1;
-    border-radius: .3rem;
-    padding: .5rem;
-    gap: .5rem;
-    display: flex;
-    align-items: center;
-    border: 1px solid rgba(0, 0, 0, .3)
-  }
-  
-  #search-bar .text-filter > input {
-    outline: none;
-    border: 0;
-    flex-grow: 1;
-  }
